token price uint256 => uint64 (#16933)
Change the token price type stored in the BridgeConfig contract from uint256 to uint64 as discussed with @longbowlu

+++ bridge/evm/test/BridgeBaseTest.t.sol
@@ -34,11 +34,11 @@ contract BridgeBaseTest is Test {
 
     address USDCWhale = 0x51eDF02152EBfb338e03E30d65C15fBf06cc9ECC;
 
-    uint256 SUI_PRICE = 12800;
-    uint256 BTC_PRICE = 432518900;
-    uint256 ETH_PRICE = 25969600;
-    uint256 USDC_PRICE = 10000;
-    uint256[] tokenPrices;
+    uint64 SUI_PRICE = 12800;
+    uint64 BTC_PRICE = 432518900;
+    uint64 ETH_PRICE = 25969600;
+    uint64 USDC_PRICE = 10000;
+    uint64[] tokenPrices;
     address[] supportedTokens;
     uint8[] supportedChains;
 
@@ -101,7 +101,7 @@ contract BridgeBaseTest is Test {
         supportedTokens[4] = USDT;
         supportedChains = new uint8[](1);
         supportedChains[0] = 0;
-        tokenPrices = new uint256[](5);
+        tokenPrices = new uint64[](5);
         tokenPrices[0] = SUI_PRICE;
         tokenPrices[1] = BTC_PRICE;
         tokenPrices[2] = ETH_PRICE;
