+++ bridge/evm/test/SuiBridgeTest.t.sol
@@ -477,7 +477,7 @@ contract SuiBridgeTest is BridgeBaseTest, ISuiBridge {
 
         committee.initialize(_committee, _stake, minStakeRequired);
         vault = new BridgeVault(wETH);
-        tokenPrices = new uint256[](5);
+        tokenPrices = new uint64[](5);
         tokenPrices[0] = 10000; // SUI PRICE
         tokenPrices[1] = 10000; // BTC PRICE
         tokenPrices[2] = 10000; // ETH PRICE
@@ -563,7 +563,7 @@ contract SuiBridgeTest is BridgeBaseTest, ISuiBridge {
         committee = new BridgeCommittee();
         committee.initialize(_committee, _stake, minStakeRequired);
         vault = new BridgeVault(wETH);
-        tokenPrices = new uint256[](5);
+        tokenPrices = new uint64[](5);
         tokenPrices[0] = 10000; // SUI PRICE
         tokenPrices[1] = 10000; // BTC PRICE
         tokenPrices[2] = 10000; // ETH PRICE
@@ -732,7 +732,7 @@ contract SuiBridgeTest is BridgeBaseTest, ISuiBridge {
         committee = new BridgeCommittee();
         committee.initialize(_committee, _stake, minStakeRequired);
         vault = new BridgeVault(wETH);
-        tokenPrices = new uint256[](5);
+        tokenPrices = new uint64[](5);
         tokenPrices[0] = 10000; // SUI PRICE
         tokenPrices[1] = 10000; // BTC PRICE
         tokenPrices[2] = 10000; // ETH PRICE
